Renamed upgrader subfolders
