+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/ExtraLifeSpawner/SpawnRateMaximumUpgrader.cs
