+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/ExtraLifeSpawner/SpawnRateMinimumUpgrader.cs
