+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/AmmoRequestsMaximumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Shared;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.JackolanternUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class AmmoRequestsMinimumUpgrader : LeafUpgrader
+    public class AmmoRequestsMaximumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.AmmoRequestsMinimum.LimitedValue += 1;
+            _jackolanternData.AmmoRequestsMaximum.LimitedValue += 1;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.AmmoRequestsMinimum.Maximum,
-            _jackolanternData.AmmoRequestsMinimum.LimitedValue);
+            _jackolanternData.AmmoRequestsMaximum.Maximum,
+            _jackolanternData.AmmoRequestsMaximum.LimitedValue);
     }
 }
