+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/AmmoRequestsMinimumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Shared;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.JackolanternUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class AmmoRequestsMaximumUpgrader : LeafUpgrader
+    public class AmmoRequestsMinimumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.AmmoRequestsMaximum.LimitedValue += 1;
+            _jackolanternData.AmmoRequestsMinimum.LimitedValue += 1;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.AmmoRequestsMaximum.Maximum,
-            _jackolanternData.AmmoRequestsMaximum.LimitedValue);
+            _jackolanternData.AmmoRequestsMinimum.Maximum,
+            _jackolanternData.AmmoRequestsMinimum.LimitedValue);
     }
 }
