+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/FireForceMaximumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Difficulty.Upgraders.CompositeCore;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class FireRateMaximumUpgrader : LeafUpgrader
+    public class FireForceMaximumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.FireRateMaximum.LimitedValue *= Constants.Difficulty.DefaultInverseLevelUpFactor;
+            _jackolanternData.FireForceMaximum.LimitedValue *= Constants.Difficulty.DefaultLevelUpFactor;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.FireRateMaximum.Minimum,
-            _jackolanternData.FireRateMaximum.LimitedValue);
+            _jackolanternData.FireForceMaximum.Maximum,
+            _jackolanternData.FireForceMaximum.LimitedValue);
     }
 }
