+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/FireForceMinimumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Difficulty.Upgraders.CompositeCore;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.JackolanternUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class FireForceMaximumUpgrader : LeafUpgrader
+    public class FireForceMinimumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.FireForceMaximum.LimitedValue *= Constants.Difficulty.DefaultLevelUpFactor;
+            _jackolanternData.FireForceMinimum.LimitedValue *= Constants.Difficulty.DefaultLevelUpFactor;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.FireForceMaximum.Maximum,
-            _jackolanternData.FireForceMaximum.LimitedValue);
+            _jackolanternData.FireForceMinimum.Maximum,
+            _jackolanternData.FireForceMinimum.LimitedValue);
     }
 }
