+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/FireRateMaximumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Difficulty.Upgraders.CompositeCore;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.JackolanternUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class FireForceMinimumUpgrader : LeafUpgrader
+    public class FireRateMaximumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.FireForceMinimum.LimitedValue *= Constants.Difficulty.DefaultLevelUpFactor;
+            _jackolanternData.FireRateMaximum.LimitedValue *= Constants.Difficulty.DefaultInverseLevelUpFactor;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.FireForceMinimum.Maximum,
-            _jackolanternData.FireForceMinimum.LimitedValue);
+            _jackolanternData.FireRateMaximum.Minimum,
+            _jackolanternData.FireRateMaximum.LimitedValue);
     }
 }
