+++ Assets/Scripts/Difficulty/Upgraders/LeafUpgraders/Jackolanterns/FireRateMinimumUpgrader.cs
@@ -4,19 +4,19 @@ using LimboOfCeres.Scripts.Shared.ScriptableObjectsDefinitions;
 using UnityEngine;
 
 
-namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders
+namespace LimboOfCeres.Scripts.Difficulty.Upgraders.LeafUpgraders.Jackolanterns
 {
-    public class AimRateMaximumUpgrader : LeafUpgrader
+    public class FireRateMinimumUpgrader : LeafUpgrader
     {
         [SerializeField] private JackolanternScriptable _jackolanternData;
 
         protected override void OnUpgradeHook()
         {
-            _jackolanternData.AimRateMaximum.LimitedValue *= Constants.Difficulty.DefaultInverseLevelUpFactor;
+            _jackolanternData.FireRateMinimum.LimitedValue *= Constants.Difficulty.DefaultInverseLevelUpFactor;
         }
 
         public override bool IsAtLimit => Mathf.Approximately(
-            _jackolanternData.AimRateMaximum.Minimum,
-            _jackolanternData.AimRateMaximum.LimitedValue);
+            _jackolanternData.FireRateMinimum.Minimum,
+            _jackolanternData.FireRateMinimum.LimitedValue);
     }
 }
